Modify: globals.css div for home
